Kinematic chain library
