+++ src/kinematic-chain.cpp
@@ -1,40 +1,48 @@
-#include <stdexcept>
+#include "serial-kinematics/kinematic-chain.hpp"
 
-#include "serial-kinematics/kinematics.hpp"
+#include <cassert>
+#include <stdexcept>
 
 namespace kinematics {
 
-using HomMat = Eigen::Matrix4d;
-using RotMat = Eigen::Matrix3d;
-using HomVec = Eigen::Vector4d;
-using PosVec = Eigen::Vector3d;
+// TODO: Formalize these assumptions in documentation:
+// 1. Frame 0 is a fixed origin Frame
+// 2. Row 0 of DH parameters corresponds to link 1, etc.
+KinematicChain::KinematicChain(const DenhavitHartenbergParam &params)
+    : _params{params}, _dof{params.dof()} {
+  _joint_vars.resize(_dof);
+}
 
-class KinematicChain {
- public:
-  KinematicChain(const DenhavitHartenbergParam &params)
-      : _params{params}, _dof{params.dof()} {
-    _joint_vars.resize(_dof);
+void KinematicChain::set_joint_vars(Eigen::VectorXd joint_vars) {
+  if (joint_vars.size() != _dof) {
+    throw std::invalid_argument("Joint variables must be _dof x 1.");
+  }
+  _joint_vars = joint_vars;
+}
+
+/**
+ * @brief Get transform from frame index ind_from to frame index ind_to, where
+ * index 0 corresponds to the fixed base frame.
+ *
+ * - By default gives transformations to the origin frame.
+ * - Works for transformations where ind_from < ind_to.
+ */
+HomMat KinematicChain::get_transform_upto(const int ind_from,
+                                          const int ind_to) {
+  if (ind_from < 0 || ind_to >= _dof || ind_to < 0 ||
+      ind_from >= _dof) {  // TODO: Check if this should be > or >=
+    throw std::range_error(
+        "Transforms outside the range [0, _dof] are inaccessible.");
   }
 
-  HomMat get_transform_upto(int ind) {
-    if (ind < 0 || ind >= _dof) {  // TODO: Check if this should be > or >=
-      throw std::range_error(
-          "Transforms outside the range [0, _dof] are inaccessible.");
-    }
-
-    HomMat out{HomMat::Identity()};
-    for (int ii{0}; ii < ind; ++ii) {  // TODO: Check what happens when ii is 0
-      out = out * get_transform(_params[ii], _joint_vars[ii]);
-    }
-
-    return out;
+  HomMat out{HomMat::Identity()};
+  int step = sgn(ind_from - ind_to);
+  for (int ii{ind_to}; ii != ind_from;
+       ii += step) {  // TODO: Check what happens when ii is 0
+    out = out * get_transform(_params[ii], _joint_vars[ii]);
   }
 
- private:
-  const DenhavitHartenbergParam _params{};
-  const int _dof{};
-
-  Eigen::VectorXd _joint_vars{};
-};
+  return out;
+}
 
 }  // namespace kinematics
